refactor : 파일 위치 변경, 메서드 이름 변경
