+++ src/main/java/com/example/nbcnewsfeed/common/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.example.nbcnewsfeed;
+package com.example.nbcnewsfeed.common.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
