+++ src/main/java/com/example/nbcnewsfeed/friend/entity/FriendRequest.java
@@ -1,6 +1,6 @@
 package com.example.nbcnewsfeed.friend.entity;
 
-import com.example.nbcnewsfeed.BaseEntity;
+import com.example.nbcnewsfeed.common.entity.BaseEntity;
 import com.example.nbcnewsfeed.user.User;
 import jakarta.persistence.*;
 import lombok.Getter;
