+++ src/main/java/com/example/nbcnewsfeed/user/FakeUserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.web.server.ResponseStatusException;
 
 @Repository
 public interface FakeUserRepository extends JpaRepository<User, Long> {
-    default User findUserByid(Long userId){
+    default User findByIdOrElseThrow(Long userId){
         return findById(userId).orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND,"User not found"));
     };
 }
